requirements.txt
flask
flask-security-too
sqlalchemy
loguru
celery